Less strict oelint.vars.multilineident
the indent on multiline variables doesn't need to exactly as long as the
variable name plus the operator sign. more important it is that there is
at least a indent (4 blanks ought to be enough) and that across all the
lines the same indent is used.
So the rule changes to reflect that.
Minimum of 4 spaces indent, same indent among all lines.
If different indention is found the indent with the highest distribution
is used as the desired one

Closes #201

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>