cls_rule: mind appendix from rule files
first of all on loading the rules and filtering them.
In this case we need to check if a rule is just defined with an
appendix and load it, otherwise ignore it for the loaded ruleset.
Secondly when computing the actual findings we need to check on
all potential combinations when a rule-file was loaded.
If there is a match we take the severity from the rule-file, otherwise
remain on standard severity.
For that we are adding a few new variables, as we need to differentiate
between internal list of possibilities and attributes for display.

Added a few test cases to avoid future regressions

Closes #236