cls_rule: Fix default severity from rulefile
When a rule's value in the rulefile is set to "", then the rule should
use its default severity.  This behavior regressed in the last release.

Properly fall back to the default severity when "" is found in the
rulefile and add a testcase to prevent this from happening again in the
future.

Fixes: #241