srcurioptions: Allow type=kmeta option
The linux-yocto class uses the `type=kmeta` SRC_URI option to designate
kernel metadata sources.  Thus, oelint-adv should not trigger a lint
about it being unknown.

As there is no good way to check for linux-yocto inheritance when
bbappends across layers are involved, the chosen solution is to always
allow this option.  The chance for a false negatives should be quite
low...

Fixes: #245