Handle single-quote quotations as valid
The oe-core dunfell branch as of commit 6cd21ddc6f ("tzdata: update 2021d -> 2021e")
has meta/recipes-bsp/u-boot/u-boot.inc include file, which contains the
following quotation:

 EXTRA_OEMAKE = 'CROSS_COMPILE=${TARGET_PREFIX} CC="${TARGET_PREFIX}gcc ${TOOLCHAIN_OPTIONS}" V=1'

This is valid and the intention is to pass the content inside of quoted
string to make without expanding the variables in the quoted string first.

Consider this single-quote syntax valid and do not emit an error in such case.

Signed-off-by: Marek Vasut <marex@denx.de>