Keep relative path to fix ValueError when sorting by filenames
ee585dca16ae07fe066c8c1154ea733e18e59a12 introduced changes that result
in an ValueError: "<filename> not in list" when sorting the filenames
at the end of the group_files function.

This reverts the resolution of f to an absolute path, which would
otherwise lead to an inconsistance as the function uses relative paths
everywhere else.