rulefile: enabled use of subids only
In the latest rework an issues slipped in that
allowed only entire groups (or ID) to be
enabled/disabled, mind the subIDs as well when
loading the rules.
As they are filtered out afterwards it should be fine.
This applies of course only to the rule itself,
thus we don't run into conflicts with the fix method call

Closes #436

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>