core: enforce fork as mp context
as we need to share all resources of the linter's
main process with the child tasks running the checks.
Otherwise things like the CONSTANTS interface will
be simply reinitialized within the child's space,
but ignoring all the added things from the tweaks
runs by the main process.

This will be mandatory from python 3.14

Closes #721
Closes #722

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>