libmode: allow non-persistent mode
that doesn't overwrite files, but instead returns
the fixed files as strings.
For this the `run` method accepts now a new
parameter `persistent` (default True).
If being set to False no file is overwritten.

`run` now also returns a tuple of the linter findings
and a dictionary of filename and fixed strings.

This allows consuming scripts to take control
of any writing behavior on their own

Closes #559

Signed-off-by: Konrad Weihmann <kweihmann@outlook.com>